feat(BreadCrumbs): add search trigger component to breadcrumbs for enhanced navigation; remove sidebar search trigger
